requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# API Clients & Web Scraping
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Templating (for content generation)
jinja2==3.1.3

# AI Curation
anthropic==0.43.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Scheduling
apscheduler==3.10.4

# Data Processing
python-dateutil==2.8.2
pytz==2024.1

# Utilities
pyyaml==6.0.1
python-slugify==8.0.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
